initialization öğrendim xd

+++ src/include/initializer.hpp
@@ -20,7 +20,7 @@ public:
   bool try_init(std::vector<cv::Point2f>& kpts_ref,
                 std::vector<cv::Point2f>& kpts_curr,
                 std::vector<int>& vector_tracked_p3d_ids_global,
-                std::vector<cv::Point3d>& vector_p3d_global);
+                std::vector<cv::Point3f>& vector_p3d_global);
